SWAGGER is used to comment out each method used in all controllers

+++ wolox/src/main/java/com/tech/wolox/service/AlbumsService.java
@@ -19,7 +19,7 @@ public interface AlbumsService {
 
     AlbumDTO[] getAlbums();
 
-    AlbumDTO[] getAlbumsbyUser(Long userId);
+    AlbumDTO[] getAlbumsbyUser(Integer userId);
     
     AlbumDTO createAlbum(AlbumDTO albumDTO);
